- Fixed merge problems - New music added - New mainmenu music - Changed "Pistol M9" ammunition size

+++ Cyber Infection/Assets/CyberInfection/Scripts/GameMechanics/Entity/Unit.cs
@@ -57,7 +57,7 @@ namespace CyberInfection.GameMechanics.Entity
 			photonView.RPC(CachedRPC.ReceiveDamage, RpcTarget.All, JsonUtility.ToJson(damageData));
 		}
 
-        public void RestoreHealth(float healthAmount)
+        public void RestoreHealth(int healthAmount)
         {
             if (PhotonNetwork.OfflineMode)
             {
